+++ packages/dev/core/src/Engines/nativeEngine.ts
@@ -253,6 +253,7 @@ export class NativeEngine extends Engine {
             highPrecisionShaderSupported: true,
             colorBufferFloat: false,
             supportFloatTexturesResolve: false,
+            rg11b10ufColorRenderable: false,
             textureFloat: true,
             textureFloatLinearFiltering: false,
             textureFloatRender: true,
@@ -280,7 +281,7 @@ export class NativeEngine extends Engine {
         };
 
         this._features = {
-            forceBitmapOverHTMLImageElement: false,
+            forceBitmapOverHTMLImageElement: true,
             supportRenderAndCopyToLodForFloatTextures: false,
             supportDepthStencilTexture: false,
             supportShadowSamplers: false,