+++ packages/dev/core/src/Misc/fileTools.ts
@@ -14,8 +14,10 @@ import { ThinEngine } from "../Engines/thinEngine";
 import { EngineStore } from "../Engines/engineStore";
 import { Logger } from "./logger";
 import { TimingTools } from "./timingTools";
+import type { INative } from "../Engines/Native/nativeInterfaces";
 
 const Base64DataUrlRegEx = new RegExp(/^data:([^,]+\/[^,]+)?;base64,/i);
+declare const _native: INative;
 
 /** @ignore */
 export class LoadFileError extends RuntimeError {
@@ -175,8 +177,9 @@ export const LoadImage = (
     mimeType: string = "",
     imageBitmapOptions?: ImageBitmapOptions
 ): Nullable<HTMLImageElement> => {
-    if (typeof HTMLImageElement === "undefined") {
-        onError("Displacement map is only supported in web environment.");
+    const engine = EngineStore.LastCreatedEngine;
+    if (typeof HTMLImageElement === "undefined" && !engine?._features.forceBitmapOverHTMLImageElement) {
+        onError("LoadImage is only supported in web or BabylonNative environments.");
         return null;
     }
 
@@ -198,8 +201,6 @@ export const LoadImage = (
         url = FileToolsOptions.PreprocessUrl(input);
     }
 
-    const engine = EngineStore.LastCreatedEngine;
-
     const onErrorHandler = (exception: any) => {
         if (onError) {
             const inputText = url || input.toString();
@@ -207,7 +208,7 @@ export const LoadImage = (
         }
     };
 
-    if (typeof Image === "undefined" || (engine?._features.forceBitmapOverHTMLImageElement ?? false)) {
+    if (engine?._features.forceBitmapOverHTMLImageElement) {
         LoadFile(
             url,
             (data) => {