+++ packages/dev/core/src/Rendering/geometryBufferRendererSceneComponent.ts
@@ -19,11 +19,16 @@ declare module "../scene" {
 
         /**
          * Enables a GeometryBufferRender and associates it with the scene
-         * @param ratio defines the scaling ratio to apply to the renderer (1 by default which means same resolution). You can also directly pass a width and height for the generated textures
+         * @param ratioOrDimensions defines the scaling ratio to apply to the renderer (1 by default which means same resolution). You can also directly pass a width and height for the generated textures
          * @param depthFormat Format of the depth texture (default: Constants.TEXTUREFORMAT_DEPTH16)
+         * @param textureTypesAndFormats The types and formats of textures to create as render targets. If not provided, all textures will be RGBA and float or half float, depending on the engine capabilities.
          * @returns the GeometryBufferRenderer
          */
-        enableGeometryBufferRenderer(ratio?: number | { width: number; height: number }, depthFormat?: number): Nullable<GeometryBufferRenderer>;
+        enableGeometryBufferRenderer(
+            ratioOrDimensions?: number | { width: number; height: number },
+            depthFormat?: number,
+            textureTypesAndFormats?: { [key: number]: { textureType: number; textureFormat: number } }
+        ): Nullable<GeometryBufferRenderer>;
 
         /**
          * Disables the GeometryBufferRender associated with the scene
@@ -47,13 +52,14 @@ Object.defineProperty(Scene.prototype, "geometryBufferRenderer", {
 
 Scene.prototype.enableGeometryBufferRenderer = function (
     ratio: number | { width: number; height: number } = 1,
-    depthFormat = Constants.TEXTUREFORMAT_DEPTH16
+    depthFormat = Constants.TEXTUREFORMAT_DEPTH16,
+    textureTypesAndFormats?: { [key: number]: { textureType: number; textureFormat: number } }
 ): Nullable<GeometryBufferRenderer> {
     if (this._geometryBufferRenderer) {
         return this._geometryBufferRenderer;
     }
 
-    this._geometryBufferRenderer = new GeometryBufferRenderer(this, ratio, depthFormat);
+    this._geometryBufferRenderer = new GeometryBufferRenderer(this, ratio, depthFormat, textureTypesAndFormats);
     if (!this._geometryBufferRenderer.isSupported) {
         this._geometryBufferRenderer = null;
     }